Update README.md
The latest update seems to have broken my dashboard. After looking at the issue it seems that renaming things made it so that the custom card name needs a `lovelace-` prefix now.

This change might not be what you intended as it will break everyones dashboards so maybe a better solution is to make it work as before.

Only reason why I'm opening a PR is because there's no issue tracker and I don't really understand lovelace enough to fix it myself.